+++ Source/FarmingHysteresis/Patch/PlantGrowers_GetGizmos.cs
@@ -53,7 +53,7 @@ internal class GetGizmosPatcher
         }
 
         var data = getHysteresisData(plantToGrowSettable);
-        var harvestedThingDef = plantToGrowSettable.GetPlantDefToGrow()?.plant.harvestedThingDef;
+        var harvestedThingDef = plantToGrowSettable.GetPlantDefToGrow()?.plant?.harvestedThingDef;
 
         var harvestHysteresisCommand = new Command_Toggle
         {